+++ lambdas/functions/webhook/src/sqs/index.ts
@@ -1,7 +1,6 @@
 import { SQS, SendMessageCommandInput } from '@aws-sdk/client-sqs';
 import { WorkflowJobEvent } from '@octokit/webhooks-types';
-import { createChildLogger, getTracedAWSV3Client } from '@terraform-aws-github-runner/aws-powertools-util';
-import { Config } from '../ConfigResolver';
+import { createChildLogger, getTracedAWSV3Client } from '@aws-github-runner/aws-powertools-util';
 
 const logger = createChildLogger('sqs');
 
@@ -49,23 +48,3 @@ export const sendActionRequest = async (message: ActionRequestMessage): Promise<
 
   await sqs.sendMessage(sqsMessage);
 };
-
-export async function sendWebhookEventToWorkflowJobQueue(message: GithubWorkflowEvent, config: Config): Promise<void> {
-  if (!config.workflowJobEventSecondaryQueue) {
-    return;
-  }
-
-  const sqs = new SQS({ region: process.env.AWS_REGION });
-  const sqsMessage: SendMessageCommandInput = {
-    QueueUrl: String(config.workflowJobEventSecondaryQueue),
-    MessageBody: JSON.stringify(message),
-  };
-
-  logger.debug(`Sending Webhook events to the workflow job queue: ${config.workflowJobEventSecondaryQueue}`);
-
-  try {
-    await sqs.sendMessage(sqsMessage);
-  } catch (e) {
-    logger.warn(`Error in sending webhook events to workflow job queue: ${(e as Error).message}`);
-  }
-}