+++ modules/multi-runner/webhook.tf
@@ -1,16 +1,16 @@
 module "webhook" {
-  source      = "../webhook"
-  prefix      = var.prefix
-  tags        = local.tags
-  kms_key_arn = var.kms_key_arn
-
+  source                              = "../webhook"
+  prefix                              = var.prefix
+  tags                                = local.tags
+  kms_key_arn                         = var.kms_key_arn
+  eventbridge                         = var.eventbridge
   runner_matcher_config               = local.runner_config
   matcher_config_parameter_store_tier = var.matcher_config_parameter_store_tier
+
   ssm_paths = {
     root    = local.ssm_root_path
     webhook = var.ssm_paths.webhook
   }
-  sqs_workflow_job_queue = length(aws_sqs_queue.webhook_events_workflow_job_queue) > 0 ? aws_sqs_queue.webhook_events_workflow_job_queue[0] : null
 
   github_app_parameters = {
     webhook_secret = module.ssm.parameters.github_app_webhook_secret