+++ modules/termination-watcher/variables.tf
@@ -1,18 +1,17 @@
 variable "config" {
   description = <<-EOF
-    Configuration for the spot termination watcher lambda function.
+    Configuration for the spot termination watcher.
 
     `aws_partition`: Partition for the base arn if not 'aws'
     `architecture`: AWS Lambda architecture. Lambda functions using Graviton processors ('arm64') tend to have better price/performance than 'x86_64' functions.
     `environment_variables`: Environment variables for the lambda.
-    `enable_metric`: Enable metric for the lambda. If `spot_warning` is set to true, the lambda will emit a metric when it detects a spot termination warning.
+    'features': Features to enable the different lambda functions to handle spot termination events.
     `lambda_principals`: Add extra principals to the role created for execution of the lambda, e.g. for local testing.
     `lambda_tags`: Map of tags that will be added to created resources. By default resources will be tagged with name and environment.
     `log_level`: Logging level for lambda logging. Valid values are  'silly', 'trace', 'debug', 'info', 'warn', 'error', 'fatal'.
     `logging_kms_key_id`: Specifies the kms key id to encrypt the logs with
     `logging_retention_in_days`: Specifies the number of days you want to retain log events for the lambda log group. Possible values are: 0, 1, 3, 5, 7, 14, 30, 60, 90, 120, 150, 180, 365, 400, 545, 731, 1827, and 3653.
     `memory_size`: Memory size linit in MB of the lambda.
-    `metrics_namespace`: Namespace for the metrics emitted by the lambda.
     `prefix`: The prefix used for naming resources.
     `role_path`: The path that will be added to the role, if not set the environment name will be used.
     `role_permissions_boundary`: Permissions boundary that will be added to the created role for the lambda.
@@ -29,19 +28,28 @@ variable "config" {
     `zip`: File location of the lambda zip file.
   EOF
   type = object({
-    aws_partition = optional(string, null)
-    architecture  = optional(string, null)
-    enable_metric = optional(object({
-      spot_warning = optional(bool, false)
-    }))
-    environment_variables     = optional(map(string), {})
+    aws_partition         = optional(string, null)
+    architecture          = optional(string, null)
+    enable_metric         = optional(string, null)
+    environment_variables = optional(map(string), {})
+    features = optional(object({
+      enable_spot_termination_handler              = optional(bool, true)
+      enable_spot_termination_notification_watcher = optional(bool, true)
+    }), {})
     lambda_tags               = optional(map(string), {})
     log_level                 = optional(string, null)
     logging_kms_key_id        = optional(string, null)
     logging_retention_in_days = optional(number, null)
     memory_size               = optional(number, null)
-    metrics_namespace         = optional(string, null)
-    prefix                    = optional(string, null)
+    metrics = optional(object({
+      enable    = optional(bool, false)
+      namespace = optional(string, "GitHub Runners")
+      metric = optional(object({
+        enable_spot_termination         = optional(bool, true)
+        enable_spot_termination_warning = optional(bool, true)
+      }), {})
+    }), {})
+    prefix = optional(string, null)
     principals = optional(list(object({
       type        = string
       identifiers = list(string)
@@ -64,4 +72,9 @@ variable "config" {
     }), {})
     zip = optional(string, null)
   })
+
+  validation {
+    condition     = var.config.enable_metric == null
+    error_message = "enable_metric is deprecated, use metrics.enable instead."
+  }
 }